+++ web-sidecar-actuator/src/main/java/software/xdev/sse/web/sidecar/actuator/ActuatorWebSecurity.java
@@ -42,6 +42,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
 
+import software.xdev.sse.web.hsts.HstsApplier;
 import software.xdev.sse.web.sidecar.actuator.config.ActuatorSecurityConfig;
 import software.xdev.sse.web.sidecar.actuator.httpsecurity.ActuatorHttpSecMCustomizerContainer;
 import software.xdev.sse.web.sidecar.actuator.metrics.ActuatorSecurityMetricsHandler;
@@ -83,9 +84,11 @@ public class ActuatorWebSecurity
 	@Order(1)
 	@SuppressWarnings("java:S4502")
 	public SecurityFilterChain configureActuator(
+		final HstsApplier hstsApplier,
 		final ActuatorHttpSecMCustomizerContainer httpSecurityMatcherCustomizerContainer,
 		final WebEndpointProperties actuatorWebEndpointProperties,
-		final HttpSecurity http) throws Exception
+		final HttpSecurity http)
+		throws Exception
 	{
 		LOG.info(
 			"Building SecurityFilterChain with {} [passwordHasher={},passwordHashCache={},metricHandlers={}]",
@@ -117,6 +120,7 @@ public class ActuatorWebSecurity
 			.httpBasic(NoErrorBasicAuthenticationEntryPoint::install)
 			// NO CSRF required as we have BasicAuth anyway
 			.csrf(AbstractHttpConfigurer::disable)
+			.headers(hstsApplier::apply)
 			.authenticationManager(new ProviderManager(this.createActuatorAuthProvider()))
 			.build();
 	}